@@ -12,13 +11,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # build by specifying --build-name and --build-number, respectively.
 # In Android, build-name is used as versionName while build-number used as versionCode.
 # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 2.0.0+15
+# In Windows, build-name is used as the major, minor, and patch parts
+# of the product and file versions while build-number is used as the build suffix.
+version: 2.1.0+16
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: '>=3.3.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,60 +31,52 @@ dependencies:
   flutter:
     sdk: flutter
 
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-
-  # https://pub.dev/packages/google_mobile_ads
-  google_mobile_ads: ^1.2.0
-
-  # https://pub.dev/packages/http
-  http: ^0.13.4
-
-  # https://pub.dev/packages/shared_preferences
-  shared_preferences: ^2.0.13
-
-  # https://pub.dev/packages/url_launcher
-  url_launcher: ^6.1.3
-
-  # https://pub.dev/packages/package_info_plus
-  package_info_plus: ^1.4.2
-
-  # https://pub.dev/packages/share_plus
-  share_plus: ^4.0.4
+  cupertino_icons: ^1.0.6
 
   # https://pub.dev/packages/fluro
-  fluro: ^2.0.3
-
-  # https://pub.dev/packages/marquee
-  marquee: ^2.2.1
-
-  # https://pub.dev/packages/location
-  location: ^4.3.0
+  fluro: ^2.0.5
 
   # https://pub.dev/packages/greek_tools
-  greek_tools: ^1.0.4
+  greek_tools: ^1.0.6
 
-  # https://pub.dev/packages/progress_indicators
-  progress_indicators: ^1.0.0
+  # https://pub.dev/packages/google_mobile_ads
+  google_mobile_ads: ^4.0.0
 
-  # https://pub.dev/packages/map_launcher
-  map_launcher: ^2.2.3
+  # https://pub.dev/packages/http
+  http: ^1.2.1
 
-  # https://pub.dev/packages/google_maps_flutter
-  google_maps_flutter: ^2.1.3
+  # https://pub.dev/packages/shared_preferences
+  shared_preferences: ^2.2.2
+
+#  # https://pub.dev/packages/package_info_plus
+#  package_info_plus: ^5.0.1
 
   # https://pub.dev/packages/provider
-  provider: ^6.0.3
+  provider: ^6.1.2
+
+  # https://pub.dev/packages/url_launcher
+  url_launcher: ^6.2.5
+
+#  # https://pub.dev/packages/share_plus
+#  share_plus: ^7.2.2
+
+  # https://pub.dev/packages/google_maps_flutter
+  google_maps_flutter: ^2.6.0
 
   # https://pub.dev/packages/google_maps
-  google_maps: ^6.2.0
+  google_maps: ^7.1.0
 
   # https://pub.dev/packages/universal_html
-  universal_html: ^2.0.8
+  universal_html: ^2.2.4
 
-  # https://pub.dev/packages/collection
-  collection: ^1.16.0
+#  # https://pub.dev/packages/map_launcher
+#  map_launcher: ^3.1.0
+
+  # https://pub.dev/packages/geolocator/install
+  geolocator: ^11.0.0
 
 dev_dependencies:
   flutter_test:
